Remove use of reflect
