+++ deepequal.go
@@ -1,0 +1,12 @@
+// +build !tiny,!js
+
+// Copyright (C) 2018 Ramesh Vyaghrapuri. All rights reserved.
+// Use of this source code is governed by a MIT-style license
+// that can be found in the LICENSE file.
+
+
+package dot
+
+import "reflect"
+
+var deepEqual = reflect.DeepEqual
