+++ tiny.go
@@ -1,0 +1,10 @@
+// +build tiny js
+
+// Copyright (C) 2018 Ramesh Vyaghrapuri. All rights reserved.
+// Use of this source code is governed by a MIT-style license
+// that can be found in the LICENSE file.
+package dot
+
+func deepEqual(i1, i2 interface{}) bool {
+	return true
+}
